services/render-app/requirements.txt
python==3.6
pythonocc-core==0.18.2
flask==1.1.1
pre-commit==1.18.3
rdflib==4.2.2
SPARQLWrapper==1.8.2
requests==2.22.0
jsonpickle==0.9.6
# pickle
